# management _sysytem/app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App Name API Documentation",
...
        "version": "1.0.0",
        "x-logo": {
            "url": "http://laravel-project-template.test/logo.png",
            "backgroundColor": "#FFFFFF",
            "altText": "LDE"
        }
    },
    "servers": [
        {
            "url": "http://laravel-project-template.test",
            "description": "Production Server"
        },
        {
            "url": "http://laravel-project-template.test",
            "description": "Staging Server"
        },
        {
            "url": "http://laravel-project-template.test/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/feedback": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Get list of Feedback",
                "description": "Returns list of Feedback",
                "operationId": "getFeedbackList",
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "data retrived successfully",
                                            "type": "string",
                                            "example": "data retrived successfully"
                                        },
                                        "data": {
                                            "description": "collection of array",
                                            "items": {
                                                "$ref": "#/components/schemas/Feedback"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found "
                    }
                }
            }
        },
        "/api/v1/feedback/{feedback}": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Get information of Feedback",
                "description": " information of Feedback",
                "operationId": "getFeedbackByKey",
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "data retrived successfully",
                                            "type": "string",
                                            "example": "data retrived successfully"
                                        },
                                        "data": {
                                            "description": "collection of array",
                                            "items": {
                                                "$ref": "#/components/schemas/Feedback"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found "
                    }
                }
            },
            "put": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Update existing Feedback",
                "description": "Returns updated feedback data",
                "operationId": "updateFeedback",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Feedback key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "data updated successfully",
                                            "type": "string",
                                            "example": "data updated successfully"
                                        },
                                        "data": {
                                            "description": "collection of array",
                                            "items": {
                                                "$ref": "#/components/schemas/Feedback"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unproccessable Entity"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Delete existing Feedback",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteFeedback",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Feedback key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/candidacies/{candidacy}/feedback": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Store new Feedback",
                "description": "Create New Feedback",
                "operationId": "storeFeedback",
                "parameters": [
                    {
                        "name": "candidacy",
                        "in": "path",
                        "description": "Key of Candidacy",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "verdict": {
                                        "type": "string",
                                        "example": "yes"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "He was good at communication skill and technical concepts were clear."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "data created successfully",
                                            "type": "string",
                                            "example": "Feedback Created Successfully"
                                        },
                                        "data": {
                                            "description": "Use to indentify record insted of id,  Mustbe in 6 latters",
                                            "type": "string",
                                            "example": "xYsw03"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "response5xx": {
                "description": "Response for 5xx type of errors",
                "properties": {
                    "success": {
                        "description": "Shows status of operation",
                        "type": "boolean",
                        "default": false
                    },
                    "message": {
                        "description": "Error message to be shown to user",
                        "type": "string",
                        "example": "Database error."
                    },
                    "error": {
                        "description": "Full error for debugging",
                        "example": {
                            "key1": "data1",
                            "key2": "data2"
                        }
                    }
                },
                "type": "object"
            },
...
            "Feedback": {
                "title": "RequestBody",
                "description": "Feedback request body data",
                "required": [
                    "feedback",
                    "description",
                    "user_id"
                ],
                "properties": {
                    "key": {
                        "description": "Key",
                        "type": "string",
                        "example": "hj!*k43j4h3j"
                    },
                    "verdict": {
                        "description": "verdict",
                        "type": "string",
                        "example": "Yes"
                    },
                    "description": {
                        "description": "description about given feedback",
                        "type": "string",
                        "example": "have good practical knowledge"
                    },
                    "user_id": {
                        "description": "id of evalutor",
                        "type": "integer",
                        "example": "hjkj!%^$^%bb"
                    },
                    "candidacy_id": {
                        "description": "candidacy_id",
                        "type": "integer",
                        "example": "hjghj!&^*JBNBJ"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Header": {
                "type": "apiKey",
                "description": "This API uses TOKEN KEY authentication\n\nKey MUST be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "header",
                "scheme": "https"
            },
            "Token": {
                "type": "apiKey",
                "description": "This API uses **TOKEN KEY** authentication\n\nKey `MUST` be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "query",
                "scheme": "https"
            }
        }
    },
...
}